+++ novobi_addons/sirena_base/__manifest__.py
@@ -1,0 +1,11 @@
+{
+ 'name': "Sirena Base Module",
+ 'summary': "Contains customized code for Sirena",
+ 'author': "Novobi LLC",
+ 'category': 'Others',
+ 'version': '14.1.0.0',
+ 'depends': ['base'],
+ 'data': [
+  'data/res_company_override.xml',
+ ],
+}
