+++ novobi_addons/sirena_product/__manifest__.py
@@ -1,0 +1,12 @@
+{
+ 'name': "Sirena Product Template",
+ 'summary': "Modifies/Adds new fields to the Product Template model",
+ 'description': "Adds some customized fields for the product.template model",
+ 'author': "Novobi LLC",
+ 'category': 'Product',
+ 'version': '14.1.0.0',
+ 'depends': ['base','product'],
+ 'data': [
+  'views/product_template_form_inherited.xml',
+ ],
+}
