+++ novobi_addons/sirena_product/views/product_template_form_inherited.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="product_template_form_inherited" model="ir.ui.view">
+        <field name="name">product.template.form.inherited</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_only_form_view"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='barcode']" position="after">
+                <field name="part_name_in_chinese" string="Part Name in Chinese"/>
+                <field name="drawing_number" string="Drawing Number"/>
+                <field name="substance" string="Substance"/>
+                <field name="parts_spec_color" string="Parts Spec / Color"/>
+                <field name="part_code" string="Part Code"/>
+            </xpath>
+            <xpath expr="//form[@name='Product Template']/sheet[1]/notebook[1]/page[@name='general_information']/group[1]/group[@name='group_standard_price']/label[1]" position="before">
+                <field name="pricelist_id"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
