--- sirena_product/models/__init__.py
@@ -1,1 +1,0 @@
-from . import product_template
